Some more restructure of modules
