+++ src/Database/Design/Ampersand/Components.hs
@@ -20,6 +20,18 @@ import Control.Monad
 import System.FilePath
 import Data.Time.Clock.POSIX
 import qualified Data.ByteString.Lazy as L
+import Data.List
+import Data.Function (on)
+
+import System.Exit
+import Database.Design.Ampersand.Output.ToJSON.ToJson  (generateJSONfiles)
+import Database.Design.Ampersand.Prototype.WriteStaticFiles   (writeStaticFiles)
+import Database.Design.Ampersand.Core.AbstractSyntaxTree
+import Database.Design.Ampersand.Prototype.GenBericht  (doGenBericht)
+import Database.Design.Ampersand.Prototype.ValidateSQL (validateRulesSQL)
+import Database.Design.Ampersand.Prototype.GenFrontend (doGenFrontend, clearTemplateDirs)
+import Database.Design.Ampersand.Prototype.ProtoUtil   (installComposerLibs)
+import Database.Design.Ampersand.ADL1.Expression (notCpl)
 
 --  | The FSpec is the datastructure that contains everything to generate the output. This monadic function
 --    takes the FSpec as its input, and spits out everything the user requested.
@@ -33,6 +45,9 @@ generateAmpersandOutput fSpec =
     ; when (genFPAExcel (getOpts fSpec)) $ doGenFPAExcel fSpec
     ; when (genPOPExcel (getOpts fSpec)) $ doGenPopsXLSX fSpec
     ; when (proofs (getOpts fSpec))      $ doGenProofs   fSpec
+    ; when (validateSQL (getOpts fSpec)) $ doValidateSQLTest fSpec
+    ; when (genPrototype (getOpts fSpec)) $ doGenProto   fSpec
+    ; when (genBericht (getOpts fSpec))  $ doGenBericht fSpec
     }
 
 doGenADL :: FSpec -> IO()
@@ -113,6 +128,76 @@ doGenPopsXLSX fSpec =
     }
    where outputFile = combine (dirOutput (getOpts fSpec)) $ replaceExtension (baseName (getOpts fSpec)++ "_generated_pop") ".xlsx"
 
-    
+doValidateSQLTest :: FSpec -> IO ()
+doValidateSQLTest fSpec =
+ do { verboseLn (getOpts fSpec) "Validating SQL expressions..."
+    ; isValidRule <- validateRulesSQL fSpec
+    ; unless isValidRule (exitWith (ExitFailure 30))
+    }
+
+doGenProto :: FSpec -> IO ()
+doGenProto fSpec =
+ do { verboseLn (getOpts fSpec) "Checking on rule violations..."
+    ; reportViolations violationsOfInvariants
+    ; reportSignals (initialConjunctSignals fSpec)
+    ; if null violationsOfInvariants || development (getOpts fSpec)
+      then do { verboseLn (getOpts fSpec) "Generating prototype..."
+              ; clearTemplateDirs fSpec
+              ; writeStaticFiles (getOpts fSpec)
+              ; generateJSONfiles fSpec
+              ; doGenFrontend fSpec
+              ; verboseLn (getOpts fSpec) "\n"
+              ; verboseLn (getOpts fSpec) $ "Prototype files have been written to " ++ dirPrototype (getOpts fSpec)
+              ; installComposerLibs fSpec
+              }
+      else do { Prelude.putStrLn $ "\nERROR: No prototype generated because of rule violations."
+                                 ++"\n(Compile with --dev to generate a prototype regardless of violations)"
+              ; exitWith $ ExitFailure 40
+              }
+    ; case testRule (getOpts fSpec) of
+             Just ruleName -> ruleTest ruleName
+             Nothing       -> return ()
+    }
+ where violationsOfInvariants :: [(Rule,[AAtomPair])]
+       violationsOfInvariants
+         = [(r,vs) |(r,vs) <- allViolations fSpec
+                   , not (isSignal r)
+           ]
+       reportViolations :: [(Rule,[AAtomPair])] -> IO()
+       reportViolations []    = verboseLn (getOpts fSpec) "No violations found."
+       reportViolations viols =
+         let ruleNamesAndViolStrings = [ (name r, showprs p) | (r,p) <- viols ]
+         in  Prelude.putStrLn $ 
+                      intercalate "\n"
+                          [ "Violations of rule "++show r++":\n"++ concatMap (\(_,p) -> "- "++ p ++"\n") rps
+                          | rps@((r,_):_) <- groupBy (on (==) fst) $ sort ruleNamesAndViolStrings
+                          ]
+
+       showprs :: [AAtomPair] -> String
+       showprs aprs = "["++intercalate ", " (map showADL aprs)++"]"
+--       showpr :: AAtomPair -> String
+--       showpr apr = "( "++(showVal.apLeft) apr++", "++(showVal.apRight) apr++" )"
+       reportSignals []        = verboseLn (getOpts fSpec) "No signals for the initial population."
+       reportSignals conjViols = verboseLn (getOpts fSpec) $ "Signals for initial population:\n" ++ intercalate "\n"
+         [   "Rule(s): "++(show . map name . rc_orgRules) conj
+         ++"\n  Conjunct   : " ++ showADL (rc_conjunct conj)
+         ++"\n  Violations : " ++ showprs viols
+         | (conj, viols) <- conjViols
+         ]
+       ruleTest :: String -> IO ()
+       ruleTest ruleName =
+        case [ rule | rule <- grules fSpec ++ vrules fSpec, name rule == ruleName ] of
+          [] -> Prelude.putStrLn $ "\nRule test error: rule "++show ruleName++" not found."
+          (rule:_) -> do { Prelude.putStrLn $ "\nContents of rule "++show ruleName++ ": "++showADL (rrexp rule)
+                         ; Prelude.putStrLn $ showContents rule
+                         ; let rExpr = rrexp rule
+                         ; let ruleComplement = rule { rrexp = notCpl (EBrk rExpr) }
+                         ; Prelude.putStrLn $ "\nViolations of "++show ruleName++" (contents of "++showADL (rrexp ruleComplement)++"):"
+                         ; Prelude.putStrLn $ showContents ruleComplement
+                         }
+        where showContents rule = "[" ++ intercalate ", " pairs ++ "]"
+                where pairs = [ "("++(show.showValADL.apLeft) v++"," ++(show.showValADL.apRight) v++")" 
+                              | (r,vs) <- allViolations fSpec, r == rule, v <- vs]
+
